requirements.txt
django==4.1.6
django-phonenumber-field[phonenumberslite]==8.0.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
Celery==5.3.1
redis==4.6.0
Celery[redis]
django_redis==5.3.0
django-cors-headers==3.13.0
requests==2.32.4
orjson==3.10.15
gunicorn==20.1.0
django-debug-toolbar==4.2.0
pillow==10.4.0